pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcus <marcus.moura@ipnet.cloud>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"
load-dotenv = "^0.1.0"
typer = {extras = ["all"], version = "^0.12.1"}
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
